% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmada-package.R
\docType{package}
\name{unmada-package}
\alias{unmada}
\alias{unmada-package}
\title{unmada: UNM Advanced Data Analysis 1 & 2}
\description{
UNM Advanced Data Analysis 1 & 2 functions and data, Stat 427/527 and 428/528.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/erikerhardt/unmada}
  \item Report bugs at \url{https://github.com/erikerhardt/unmada/issues}
}

}
\author{
\strong{Maintainer}: Erik Erhardt \email{erik@statacumen.com}

}
\keyword{internal}
